Add App file and statementHtml regression test
