+++ src/clash.rs
@@ -1,4 +1,5 @@
 use serde::{Serialize, Deserialize};
+mod app;
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -56,5 +57,29 @@ impl Clash {
         println!("Constraints:\n{}\n", cdata.constraints);
         println!("Input:\n{}\n", cdata.input_description);
         println!("Output:\n{}\n", cdata.output_description);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use app::App;
+    use directories::ProjectDirs;
+
+    fn app() -> App {
+        let project_dirs = ProjectDirs::from("com", "Clash CLI", "clash")
+            .expect("Unable to find project directory");
+        App::new(project_dirs.data_dir())
+    }
+
+    fn decimal_fraction_percentage_clash() -> String {
+        let clash_file = app().clash_dir.join("58951a69f66d23586be8084cb0969c637b07.json");
+        std::fs::read_to_string(clash_file)
+            .expect("Cannot find test for clash. Please run status command.")
+    }
+
+    #[test]
+    fn do_not_panic_when_missing_statement_html() {
+        serde_json::from_str::<Clash>(&decimal_fraction_percentage_clash()).unwrap();
     }
 }
